Ajout initial des fichiers du site web
This site is an online portal for an organization [Youth empowerment organization], which aims to present to members and visitors their mission and the different activities they organize. The site will include a home page that will present general information about the organization and its programs, a resources section through which members and visitors will find relevant links which will help them find additional information and a contact section which will allow them to contact the organization.